Consolidate all MSRs and operations on them into one file
